Reconnect delays as floats

+++ custom_components/solaredge_modbus_multi/__init__.py
@@ -45,8 +45,8 @@ CONFIG_SCHEMA = vol.Schema(
                 "modbus": vol.Schema(
                     {
                         vol.Optional("timeout"): vol.Coerce(int),
-                        vol.Optional("reconnect_delay"): vol.Coerce(int),
-                        vol.Optional("reconnect_delay_max"): vol.Coerce(int),
+                        vol.Optional("reconnect_delay"): vol.Coerce(float),
+                        vol.Optional("reconnect_delay_max"): vol.Coerce(float),
                     }
                 ),
             }
